move compileStanProgram.ts
